move patching to extension type definitions
