+++ pandas/compat/__init__.py
@@ -29,6 +29,7 @@ from pandas.compat.pyarrow import (
     pa_version_under11p0,
     pa_version_under13p0,
     pa_version_under14p0,
+    pa_version_under14p1,
 )
 
 if TYPE_CHECKING:
