+++ pandas/core/arrays/arrow/extension_types.py
@@ -5,6 +5,8 @@ from typing import TYPE_CHECKING
 
 import pyarrow
 
+from pandas.compat import pa_version_under14p1
+
 from pandas.core.dtypes.dtypes import (
     IntervalDtype,
     PeriodDtype,
@@ -112,3 +114,40 @@ class ArrowIntervalType(pyarrow.ExtensionType):
 # register the type with a dummy instance
 _interval_type = ArrowIntervalType(pyarrow.int64(), "left")
 pyarrow.register_extension_type(_interval_type)
+
+
+def patch_pyarrow():
+    # starting from pyarrow 14.0.1, it has its own mechanism
+    if not pa_version_under14p1:
+        return
+
+    # if https://github.com/pitrou/pyarrow-hotfix was installed and enabled
+    if getattr(pyarrow, "_hotfix_installed", False):
+        return
+
+    class ForbiddenExtensionType(pyarrow.ExtensionType):
+        def __arrow_ext_serialize__(self):
+            return b""
+
+        @classmethod
+        def __arrow_ext_deserialize__(cls, storage_type, serialized):
+            import io
+            import pickletools
+
+            out = io.StringIO()
+            pickletools.dis(serialized, out)
+            raise RuntimeError(
+                "forbidden deserialization of 'arrow.py_extension_type': "
+                "storage_type = %s, serialized = %r, "
+                "pickle disassembly:\n%s" % (storage_type, serialized, out.getvalue())
+            )
+
+    pyarrow.unregister_extension_type("arrow.py_extension_type")
+    pyarrow.register_extension_type(
+        ForbiddenExtensionType(pyarrow.null(), "arrow.py_extension_type")
+    )
+
+    pyarrow._hotfix_installed = True
+
+
+patch_pyarrow()
